requirements.txt
albumentations
loguru
mxnet
numpy
opencv_python
opencv_python_headless
pandas
torch
torchvision
tqdm
ptflops
python-box
python-dotenv
PyTurboJPEG
tensorboard
timm==0.6.13